subtree: use contains_shallow_copy() in ctx.files()
Summary:
For subtree commits, we only need to compute the file list if it contains shallow copy. So this diffs
check that to determine if the manifest diff is needed. Later, we will make `self._changeset.files`
an empty list if the commit contains a shallow copy, and in long-term, we will remove the file list
from commit text.

Reviewed By: muirdm

Differential Revision: D65462388

fbshipit-source-id: c85e8fa8a1d45807dd100d39990e8ea0b828eceb