Pass in current test plan from ISL client
Summary: When auto-generating a diff test plan, the current test plan is passed in from the ISL client, as it will later be used by the ISL server. On the other hand, when auto-generating a diff summary, the current test plan is passed in as `undefined`, as it is not needed to be used. Thus, this saves the load of passing in unnecessary content.

Reviewed By: evangrayk

Differential Revision: D65456150

fbshipit-source-id: 69c0dbd35b14042fc7de0aff349274e5dbc1e1e3