Add a test to demonstrate a bug with overrides
Summary:
## This stack

D64536615 added a default for `dai_general_store`. After that I removed the JK override for `dai_general_store`.

What I expected:

* `dai_general_store` = 100K

What happened instead:

* `dai_general_store` = 1

This happens because `dai_general_store` does have a `master` bookmark, it just happens to contain only one commit. So setting a default for it does nothing, we need to set an override.

While testing that config change I noticed it was not honored, which led me to finding this bug.

## This diff

Add a failing test before fixing.

Reviewed By: clara-9

Differential Revision: D65477024

fbshipit-source-id: dd0fdbbad2bff0a61da232456ad19e9bfdb6eb07