Fix a bug with overrides
Summary:
## This stack

D64536615 added a default for `dai_general_store`. After that I removed the JK override for `dai_general_store`.

What I expected:

* `dai_general_store` = 100K

What happened instead:

* `dai_general_store` = 1

This happens because `dai_general_store` does have a `master` bookmark, it just happens to contain only one commit. So setting a default for it does nothing, we need to set an override.

While testing that config change (see D65477328) I noticed it was not honored, which led me to finding this bug.

## This diff

Fix the actual bug.

Because the bug simply causes the server to ignore this config, we can land this and D65477328 in any order.

Reviewed By: clara-9

Differential Revision: D65477354

fbshipit-source-id: 7691f8afffc150d1b9783384f4e8d001cf2f0f28