megarepo_async_requests_worker -> async_requests_worker in tests
Summary:
## This stack

Final cleanup now that the "new" worker (aka mononoke/sync_requests_worker) is in production and the old one (aka mononoke/megarepo_async_requests_worker) has been shut down.

## This diff

We are using a hardcoded string to represent the worker when not running in TW, i.e. for tests. Update that.

Reviewed By: YousefSalama

Differential Revision: D65531202

fbshipit-source-id: 08d99310e10f91cd45dda37f79508592a9f7d761