Pass in summary from ISL client
Summary: When auto-generating a diff test plan, the summary is passed in from the ISL client, as it will later be used by the ISL server. On the other hand, when auto-generating a diff summary, the summary is passed in as `undefined`, as it is not needed to be used. Thus, this saves the load of passing in unnecessary content.

Reviewed By: LEON-MING

Differential Revision: D65539413

fbshipit-source-id: aa06cf265811910bfcfddaf74221ed49c4d71e7f