Return actual filteredRootId instead of parsedRootId when using getGlobFiles
Summary:
# Bug Description
When creating the remote glob object in FilteredFS, we formerly returned the parsedRootId instead of the filteredRootId. This caused issues down the line when rendering the rootId since it no longer contained a length specifier, causing undefined behavior and a crash.

# This Diff
UIse filteredRootId in the GetGlobFiles Result object instead of the parsedRootId

Reviewed By: MichaelCuevas

Differential Revision: D65488953

fbshipit-source-id: 5ab826be3c4f2ae1290d327c7ff5858d9a49196c