bookmark_service: add sharding boilerplate
Summary:
The bookmark service is currently not sharded which is fine as it's only used for hg repos. If we want to start using it for the git repos too then loading all repos is going to cause tasks to crash. `BookmarkServiceRepo` might seem like a lightweight repo but it depends on a lot of facets transitively.

This diff adds the boilerplate for making the bookmark service use shardmanager.

Reviewed By: lmvasquezg

Differential Revision: D65605644

fbshipit-source-id: 0c964abae2a6678d8c26fa91e590a775794fbee2