content_manifest_derivation: implement derivation of content manifests
Summary:
Introduce `ContentManifest`. This is the replacement for fsnodes that is sharded, and simplified to exclude additional roll-up data like counts and hashes.  This will make individual manifest blobs smaller so that they are always cacheable.  Additional data like counts and hashes will be added in secondary manifest types later on, so that we can completely replace fsnodes.

This diff implements derivation of content manifests.

Reviewed By: andreacampi

Differential Revision: D55075372

fbshipit-source-id: 1757b8eb891788312b4367ed6bb78cdccfe8e5da