Update globFiles's offloading logic
Summary:
# Context
Watchman is a primary vector into the eden globFiles endpoint. How watchman interprets and sends arguments to eden different based on how the query is structured. This change is intended to make a common query pathway more efficient(referred to as SimpleSuffix from here).
Query format example:
["query","<root>",{relative_root:"<rel_root>",fields:["content.sha1hex","name"],expression:["allof",["type","f"],["suffix",["<types>"]]]}]
This is currently interpreted as [rel_root/**], then post-filtered in watchman.
This change is intended to change this to use eden glob's searchRoot field and internal filtering.

# This Diff
Updates globFiles to only offload a specifc query. This query is a known to be slow, and also is able to be improved by being offloaded.

# Technical Details

Restrict offloading to queries matching the following watchman input.
Enable relative root matching only for html and www/html

Reviewed By: jdelliot

Differential Revision: D65302346

fbshipit-source-id: 67eba8a41a710ad81a01b9aa50958b3e40c70653