Explicitly handle undefined source/code
Summary: We had a weird code path that treated any signal with undefined source or code as being blocking. This is wrong, and it should go through the allowlist like everything else. We just need to rewrite these into a string so our allowlist can actually express how to deal with it.

Reviewed By: madgen

Differential Revision: D65635446

fbshipit-source-id: 2cdbe0d3defe6d6d63c636dd25af9c26eeeb7759