rust remove: windows remove config entry
Summary:
Remove a repo on Windows: remove the repo entry from the config.json file.

Rust "Path" create will add "\\?\" prefix to indicate the path string is in unicode. In config.json file, the keys are without such prefix so we need an extra step for Windows to strip that prefix before finding the key from config.json.

Reviewed By: muirdm

Differential Revision: D65407083

fbshipit-source-id: 2b4c928ffe3ea0d37495009cf9da7e5c493b337b