rust remove: Windows cleanup
Summary:
## Background

Remove left-over files on Windows after unmounting the Eden mount.

The files are left because of the nature of PrjFS. All the materialized files will be left after unmount so we have to manually remove them.

## Tech Details

I could not use "std::fs::remove_dir_all" because it kept giving me os error 369. I tried to do retry with a short sleep but it didn't help. That's the reason instead of relying on Rust functions to remove the files, we explicitly spawn a PowerShell command to do the work.

Reviewed By: muirdm

Differential Revision: D65520457

fbshipit-source-id: 1d1f4e1ef29d1c11e34a47ab9a2b95a1186bae80