Remove eden doctor periodic runs from eden server
Summary:
**Eden Doctor Periodic Runs Migration**
As discussed with the Eden team, we have decided to move periodic Eden doctor runs from the Eden server to a scheduled task via Windows Scheduled Tasks. This change aims to improve the efficiency and reliability of the Eden doctor checks.

**Part 2**: **Removing code changes to run eden doctor within eden server**
This diff is the second step in removing the periodic Eden doctor runs from the Eden daemon. We are reverting the code changes that previously created eden doctor run tasks within the Eden server.

**Note**: We have already implemented the necessary changes to the CPE_Eden/windows.rb chef recipe to schedule the Eden Doctor run via **Windows Scheduled Tasks - D64994193**.

Reviewed By: jdelliot

Differential Revision: D65694970

fbshipit-source-id: e1aa76327e34d7f0b0ed606de3619b536e80319f