Pass in Phabricator diff number from ISL client
Summary:
When auto-generating a diff test plan, the Phabricator diff number is passed in from the ISL client, as it will later be used by the ISL server. If there is no associated Phabricator diff number, then it is passed in as an empty string. This is so that just the `string` type can be used for ease, such as in D65564922, where the `string` type can be used directly for the parameter of `RevisionUtils::getDiffFromString`.

On the other hand, when auto-generating a diff summary, the Phabricator diff number is passed in as `undefined`, as it is not needed to be used. Thus, this saves the load of passing in unnecessary content.

Reviewed By: evangrayk

Differential Revision: D65570199

fbshipit-source-id: c840f702edf0d6c54bcdb144aa255d7f46c7dc70