renderdag: decouple tests from `dag`
Summary:
Previously, `renderdag`, `dag` form a cycle:

  `renderdag (test)` dev-dependency
    \_ `dag` optional dependency
         \_ `renderdag`

While this does not break compile, since `renderdag (test)` and `renderdag` are
two separate crates, it is problematic for `cargo publish`, which cannot handle
this case.

Although it's possible to patch `cargo publish` to skip some checks to publish,
let's just rewrite the `renderdag` tests so it does not depend on `dag`. This
way, regular `cargo publish` can handle it just fine.

Reviewed By: sggutier

Differential Revision: D65793821

fbshipit-source-id: 92ac3d246faed18e0b36ac551644b07921f7aa87