wait for requests in-flight to finish
Summary:
We want:
* to shut down quickly if there are no in flight requests rather than wait the full grace period duration. This is only getting more important with more tasks we have in region as it increases the time it takes to restart a while region which we rely on for deployments and sev mitigations.
* to not shut down dropping requests. We don't want to unnecessarily drop requests if they have a chance to finish.

In order to satisfy **both**, we need to monitor in-flight requests. The convenience function `run_until_terminated` has added `Arc<AtomicI64>` which a service can provide to achieve the both bulletpoints above.

With this diff we shut down immediatelly when there are no requests in-flight. If they are, we wait until the grace time is exhauseted.

This is especially needed for git where requests can be long-running. We can wait long time before shutting down, but we don't want to do that when we don't have to. We currently can drop requests because we're shutting down too early: https://fburl.com/scuba/dev_command_timers/k61wrmrj

Reviewed By: mitrandir77

Differential Revision: D65821465

fbshipit-source-id: da7e7b9ebfe112717b0521b0689d1fb1f38f4c67