Add mount point parameter to eden notify get-position
Summary:
# Context

We are introducing EdenFS notifications to support scalable and ergonomic file system notifications for EdenFS mounts.

# This Diff

Previious version of `eden notify get-position` only reported the position of repo located in the cwd. This diff adds the ability to specify a mount point.

# Technical Details

Added a single `mount_point` parameter to the `get-position` paramter. This paramter is positional (no flag).

# Next Steps

Add the `changes-since` command to `eden notify` then iterate on the `streamChangesSinceV2` Thrift endpoint.

# Discussion Points

None

Differential Revision: D65781302

fbshipit-source-id: 637d547c38a010564a48e140e39bab07a5199655