Update lock test to use eden_repo
Summary:
# Context
This test is a copy of lock_test but using eden_repo. The state of the these tests are currently not correct behavior.

# Technical Details

I chose to make an eden_repo specific copy rather than updating lock test due to the current differences in behavior between the eden_repo and backing repo making a unified test difficult. Ideally they would have the same results for the same inputs but currently do not.

Reviewed By: MichaelCuevas

Differential Revision: D64720773

fbshipit-source-id: f0954210fb24b5b23c286c9e71a933fd050744b6