slapi: defend against incomplete server response for bookmarks
Summary:
Sometimes the request:

  api.bookmarks(['master'])

got a HTTP 200 response that decodes to:

  {}

In some code paths, this is then interpreted as `{'master': None}`, meaning the
deletion of the `master` bookmark. That's not ideal. So let's defend against
the error by checking the response length.

Normally, the server reports all bookmark names as requested with values
being either `None` or the commit hash.

Reviewed By: muirdm

Differential Revision: D65855783

fbshipit-source-id: 304ba8e5d72bfe52fd215cf67002a4f11f9005b4