Check for invalid certs
Summary:
This diff is part 4 of the stacked changes to introduce a new health-report command in Eden. Specifically, this diff focuses on adding checks to notify if certs are invalid.

### Eden Health Check Subcommand
**Problem Statement**
Currently, when Eden runs within VSCode, we are unable to effectively surface Eden-related issues/failures to users who are unaware that Eden is not healthy on their machine. This can lead to users being unable to use Eden-dependent services like Myles and Sapling.

**Solution**
To address this issue, we propose creating a new subcommand that performs critical checks related to Eden health. This subcommand will be lightweight and can be automated via VSCode periodic runs.

**Critical Checks**
The following critical checks should be included in the subcommand:
- Eden Daemon Not Running: Check if the Eden daemon is running using eden run.
- No Repo Mount Found: Check if any repositories are mounted and if the user is actively developing using Eden-checked out repos.
- Corp/Prod Canal Certificates Missing/Expired: Check if Corp/Prod canal certificates are missing or expired, which would prevent Eden from operating.
- Eden Stale Version: Check if the Eden version is older than 28 days from the current deployed or installed version.
- Last Chef Age/Run: Check if Chef has run recently (ideally every 30 minutes) to ensure that the latest Eden version is installed.

**Benefits**
By implementing this subcommand, we can:
- Surface Eden-related issues/failures to users more effectively
- Improve user engagement with Eden-dependent services like Myles and Sapling
- Reduce the likelihood of users blaming other systems for issues caused by Eden
- Ensure that users have the latest version of Eden installed and running on their machines

Reviewed By: lXXXw

Differential Revision: D65494875

fbshipit-source-id: fddb5cc7032ea7fcd347568d342fba35653f9103