Index map of funnel trackers by fieldNameAndHashKey rather than hashKey
Summary: The map of funnel trackers is indexed by `fieldNameAndHashKey` rather than `hashKey`. The associated changes are also made, such as the dependency arrays of React Hooks, the key type of the map of funnel trackers to reflect `fieldNameAndHashKey`'s `string` type, variable and function naming, and comments. Thus, by using distinct keys, the funnel trackers and suggestion identifiers are different when auto-generating a diff summary and test plan, one after the other.

Reviewed By: evangrayk

Differential Revision: D65883920

fbshipit-source-id: 3a85f03abf49732a60f90fbefa803755a00661de