derived_data_metrics_tailer: split metrics by deriving process
Summary:
We would like to have a better understanding of which process is deriving commits.  This will let us improve our headline metrics to exclude background processes like backfilling.

We add the concept of a "deriver", which is a category of binary (determined by build rule) that performs derivation.  We split the ODS metric into subcategories.

For seamless migration, we log an aggregate stat also.

Reviewed By: mitrandir77

Differential Revision: D65839025

fbshipit-source-id: 08f584b8adab282f3d47de7fe0826b5f95f115b2