Upgrade pingora to 0.4.0
Summary:
pingora has merged the pending pull request and cut a new release 0.4.0. Use that instead of depending on the old git hash.

The new release introduces rustls which will be useful since we don't need to depend on libssl which does not work inside VMs. However, building rustls fails on some architectures. Since there is no immediate use of tls in pingora at the moment, I will leave it out. Once it's needed, we will just need to add pingora-rustls to Cargo.toml and figure out a why to make build work.

Reviewed By: chengxiong-ruan

Differential Revision: D65950550

fbshipit-source-id: 8ffec0f3baf38b7cc62bf9b5cfb7653adca9d5e1