filemerge: don't save backup file for in-memory merge
Summary:
It doesn't seem like we need to save a backup file for an in-memory file merge. The merge is all-or-nothing with no user input, so there should be no need for a backup.

I cleaned up some now defunct in-memory related code in `_makebackup()`. Previously the backup was indirectly needed after the internal merge tools were ran the premerge (to undo the failed premerge), but in D65841341 I removed the premerge for (most) internal tools.

This backup file activity showed up when profiling a large rebase with a lot of file merges.

Reviewed By: zzl0

Differential Revision: D65803891

fbshipit-source-id: 9dd2f7194d94b63ea11ff4679d1f6b8297db6775