filemerge: turn off fullmerge for most built-in tools
Summary:
"fullmerge" means to first try a "premerge" which is a normal 3-way merge. If the premerge is successful, the merge tool's specific merging is skipped. The main point (I think) is to avoid delegating to an external merge tool if a simple 3-way merge can figure things out.

Our built-in tools (e.g. ":merge") were set to "fullmerge" which means they did a premerge. This isn't necessary, though, since the internal tool itself also just does a 3-way merge (with additional logic on how to handle conflicts). It resulted in running the merge process twice in certain cases when a single time would do, among some other unnecessary work.

This change re-ordered some output (as seen in tests), but I think the new output order is an improvement.

Reviewed By: zzl0

Differential Revision: D65841341

fbshipit-source-id: 7d5ea1bbd4dadc1eaac629af2c50891509f648fc