merged all suggested change types to shared object
Summary:
Context: We currently have suggested changes coming from Human, Metamate, and CI Signal. It is more ideal to reuse the same SuggestedChange type for all of them instead of making a new field for each.

This way it is more scalable later on when the team want to add in other CodePatch types (which they plan to).

Solution: going to refactor and merge them into 1 SuggestedChange.

PS: Will bring in the fields needed for CI Signal in the next diff

Differential Revision: D65942019

fbshipit-source-id: f8bf9224fb8be509f4663edc657951f2ad64a4bb