Log when SCS methods take long to execute
Summary:
## This stack

I am trying to debug a case of some methods stalling the entire reactor. See https://fb.workplace.com/groups/1708850869939124/permalink/1836334797190730/ for context.

## This diff

Optionally log slow methods that take more than 500ms to run.

This is *not* the same thing that I'm trying to debug: a method can take as long as it needs, as long as it yields appropriately. But this is still a useful tool to have available.

Reviewed By: markbt

Differential Revision: D65942700

fbshipit-source-id: 8695892a0b23041c80d9745fe419eadbb3cf6aa8