remotefilelogctx: add fast path filenode comparison
Summary:
If both file contexts have a filenode and the filenodes match, we can be certain the file content matches.

Not sure this makes much of a difference in practice, but I keep suspecting it is causing issues, so let's just add it so I can stop worrying about it.

Reviewed By: zzl0

Differential Revision: D65973268

fbshipit-source-id: cf50f6d569ee530a6b654dd4759b7b9ccb13dbe6