mergedriver: don't reload during in-memory rebase
Summary:
Add some hacky caching to avoid reloading the merge driver code from disk for every commit being rebased.

Basically, after loading merge driver during an in-memory rebase, we cache the merge driver entry point in a global variable for the next commit to use. We only cache the "preprocess" step because in-memory merge should not need the other "conclude" step.

We also cache the starting set of all loaded modules so that we can unload the merge driver later, if needed. For example, transitioning from in-memory merge to on-disk merge should stop using the cached merge driver (because the merge driver needs to be read from disk each time).

Note that the in-memory merge driver loading loads the merge driver from disk (not from the in-memory wctx). This means that merge driver changes within the stack being rebased in-memory will _not_ take effect. Perhaps they should, but that would be hard to do (and would make this optimization impossible ;)).

Reviewed By: sggutier

Differential Revision: D65803896

fbshipit-source-id: bf42b37070d07a7ccb6eb0410603b5025c2db798