merge: avoid writing conflicts to disk for IMM
Summary:
During in-memory merge, we would previously write out to-be-merged file contents to disk and then load them back in when resolving. For in-memory merge, everything is done one-shot with no cross-process persistence needed. Optimize things a bit to store to-be-merged file contents in memory instead of futzing with the disk.

I invented a new `detacheddata()` method of filectx. This method attempts to maintain the laziness of the underlying data fetching while "detaching" or disassociating from the parent change context. This avoids lazy bytes stored in memory from being changed as other parts of the merge process make mutations to `wctx`. For example, with `detachedbytes` it is safe to do:

```
bytes = wctx[path].detachedbytes()
wctx.remove(path)

# use `bytes`
```

Reviewed By: zzl0

Differential Revision: D65803890

fbshipit-source-id: c048c907d3d6e7f55127f00015072b7bb8569b35