merge: don't write out merge state for in-memory merge
Summary:
In-memory merge is an all-or-nothing, one-shot operation, so it should never need to persist merge state between different processes. Let's skip writing out the merge state.

This has the side effect of not leaving the user in a merge state when ctrl-c'ing an in-memory rebase. I think this is desirable.

Reviewed By: liubov-dmitrieva

Differential Revision: D65803902

fbshipit-source-id: 27ea418ea0d3d6143afdf7541d1a6e7c2ec365c9