Improve eden notify changes-since reporting
Summary:
# Context

We are introducing EdenFS notifications to support scalable and ergonomic file system notifications for EdenFS mounts.

# This Diff

Previouls diffs where just using debug output. This diff improves the output to be human readable.

Note: the changes are currently returned from most recent to the provided position. In other words, in reverse.

# Next Steps

Iterate on the implementation of streamChangesSinceV2 to use ForEachDelta.

# Discussion Points

None

Reviewed By: lXXXw

Differential Revision: D65905097

fbshipit-source-id: 322c69af3055ba3e22833d5b7c4bfec0d4db92dd