Prepare to move changes since API from streaming to non-streaming
Summary:
# Context

We are introducing EdenFS notifications to support scalable and ergonomic file system notifications for EdenFS mounts.

# This Diff

Based on user feedback as well as discussions with former Watchman/EdenFS SWEs, we are planning to move from a streaming endpoint to a non-streaming endpoint. There are many reasons, but ergonomics and peformance (for this use case) are top reasons.

This diff moves some defintions from the streaming API to the non-streaming API.

# Next Steps

* Continue migrating the streaming interface to a non-streaming interface.
* Collapsing journal entries
* .t tests, Python integration tests, C++ unit tests

# Discussion Points

None

Differential Revision: D66026081

fbshipit-source-id: c7d1652f3a42d2854db5fba948706fe316376ef9