progress: show the last progress bar from Sapling on ISL
Summary:
This makes showing all progress bars from Sapling (including but not limited to` sl goto`) possible. It leverages some existing code for rendering progress bars that was previously used for jellyfish.

It is very common for Sapling to show more than one progress bar, so instead of choosing carefully between all progress bars or rendering them all nicely we just show the last one. That will be done in later diffs, as well as adding percentages in progress bars.

Reviewed By: evangrayk

Differential Revision: D65769371

fbshipit-source-id: de0d0c045c5bca60873ece4dd7bda4e1d85922de