Refactored Thrift types for eden CLI
Summary:
# Context

We are introducing EdenFS notifications to support scalable and ergonomic file system notifications for EdenFS mounts.

# This Diff

This diff creates local types (as opposed to Thrift types) to make it possible to add various traits. It also implements Display and From for these types and then updates changes-since and get-postion to use the new types

# Next Steps

* Collapsing journal entries
* .t tests, Python integration tests, C++ unit tests

# Discussion Points

None

Differential Revision: D66045966

fbshipit-source-id: 999b27c47877c2e1ced3febe983be5fee9ca1e36