only use one insetProvider
Summary:
### Context
We want to show all comments from all active repos. So both webviews don't really care about which repo the comment is from. Let's abstract this inside of the ControllerList. IMO the ControllerList would be the data abstraction for the providers. The provders shouldn't need to deal with any individual 'controller' or whatever we call it.

### This diff
- One insetProvider per repo -> one insetProvider for all repos.
- Also I'm in the process to move most if not all controller logic into the controller.

Differential Revision: D66045875

fbshipit-source-id: 54069a28fbf4fa3071444ab64ea89ed5bebc00d0