PrivHelper: Respect soft/hard mount choice
Summary:
# This diff
Teaches PrivHelper to respect the soft/hard mount choice that's passed into the nfsMount() request. We were previously ignoring that option altogether. Now we actually modify the mount command so that the correct option is chosen based on the active EdenConfig.

# Context

NFS Mounts can either be mounted as "hard mounts" or "soft mounts". The difference between these two is as follows:

* Hard mounts: If the server becomes unresponsive while an NFS file system is mounted, any new or outstanding file operations on that file system will hang uninterruptibly until the server comes back (or that NFS file system is forcibly unmounted).
* Soft mounts: If the server becomes unresponsive while an NFS file system is mounted, file system calls will fail after `retrans` round trip timeout intervals (i.e. requests will not hang until the server comes back online).

There are many arguments online about whether hard or soft mounts should be utilized. The main argument against hard mounts is that they can cause the system to hang for an indeterminate amount of time.

On macOS, some users have seen hanging mount that cause processes to be unkillable. They have also seen cases where `diskutil` (even `diskutil list`) operations hang forever. This sounds like a classic case of a hard mount hanging and any request made to that mount also infinitely hanging.

These types of issues are somewhat solved with soft mounts, but soft mounts come with their own set of issues. We will have to weigh the pros and cons of hard vs soft mounts, and to do so we will have to be able to easily try both mount types. We can allow this by adding an Eden config option that allows switching between the two types.

Reviewed By: jdelliot

Differential Revision: D65691305

fbshipit-source-id: a1821721598c9154c93149ac280f2d03f15d15d9