Update integration test before supporting multiple bookmarks
Summary:
## This stack
A few minor updates to cross-repo codebase to facilitate the syncing/catching up of multiple (hundreds) of bookmarks of cross synced repos.

This is mostly about properly supporting selecting changesets (to sync of verify) using bookmark regexes.

## This diff
Extends an integration test to prepare for the changes in the next diffs.

Reviewed By: mitrandir77

Differential Revision: D65996106

fbshipit-source-id: 2b2d3e466d15e1e989869a61394534141fda9db8