modern sync: stream-ify
Summary:
This way it'll be easier to do concurrent stuff and play with chunks later.

Right now this has the fundamental flaw that since it flattens every bookmark move into a stream of commits we can't keep track of which precise bookmark move we're syncing given we're doing it in chunks. I want to play with this a little bit and see if there'a a way of using this stream to upload content in the background before we start syncing bonsais which is when we ultimately update the counter. For now, putting this up as a POC.

Reviewed By: YousefSalama

Differential Revision: D66103472

fbshipit-source-id: c261ed2f883531b666372ceaecdf491ab1d609f3