rust-remove: user prompt and messaging
Summary:
## Background

We are migrating `eden remove` command from Python to Rust for better error handling and code readability.

## This Diff

Implemented helper functions to show user information and ask for confirmation:
- `prompt_user`: for confirmation before we perform any destructive operations.
- `display_msg`: show information that we hope user be aware of. It supports different message level which will be printed in different colors.

Also in this diff we introduced a new command line argument "quiet" which is by default hidden from the end user. This is convenient for testing when we don't care about the output (messaging) for the end user.

Reviewed By: muirdm

Differential Revision: D65766009

fbshipit-source-id: 30feb200799111a791a09b73f0209c2ced38a45b