rust-remove: remove a regular file
Summary:
## Background

We are migrating `eden remove` command from Python to Rust for better error handling and code readability.

## This Diff

Handles the case when the path provided by the user to remove is actually a file. To handle this, we prompt the user for confirmation and simply go deleting it.

Also in this diff, we introduced a testing-only flag "-n". This is because currently dialoguer::Confirm does not accept input from non-terminal.

Passing "-y" and "-n" at the same time is undefined and will cause error.

Reviewed By: muirdm

Differential Revision: D65801635

fbshipit-source-id: 8990a161936d0c132c97fe751562a90f319be30f