rust-remove: Validation state
Summary:
## Background

We are migrating `eden remove` command from Python to Rust for better error handling and code readability.

## This Diff

Implemented the `Validation` state for the state machine design of `remove` command. We added this state as a final check to make sure the ultimate state matches what the end user should expect. If this check fails, the command will be considered unsuccessful.

In the future, we might also use this state to help the user to mitigate issues. e.g. what should the user do when the directory somehow still exists.

Reviewed By: muirdm

Differential Revision: D65801370

fbshipit-source-id: 82ad9579bc111bb618c47375a288d3b7616a1ee7