hooks: Restore lost scuba logging of changeset
Summary:
In D64245747, while moving code around, we lost logging of the changeset hash when the hook isn't bypassed.

Restore it.

This task was more of a PITA than planned until I realized that the scuba logging was spread between 2 places and I needed to write code in one while being aware of the other.

Reviewed By: mzr

Differential Revision: D66177243

fbshipit-source-id: 180618ac31315f335c7271baf21c73085b348f95