added shortcut to open inline comment side bar
Summary:
Context: we want the feature to be visible for users; thus we are adding a shortcut to open the side bar

In a later diff I'll add a tool tip on the inset view to tell users about the shortcut.

PS: I just used CMD-Shift-G for now, going to change this later.

Reason being, G is to open the Git/source control tab.
HOWEVER, it is already used for "FIND PREVIOUS", but who uses find previous when you can just Cmd-F and shift/noshift + enter to go previous and next?

But yea I just put it as this for now and will change later.

I also need to see what keybinds are available on windows machine.

Differential Revision: D66211796

fbshipit-source-id: 4446a0da6477d427d0f69003c0cb9d8ae172fe56