histedit: simplify computation of new list of files
Summary:
Similar to treatment given to "amend" in previous commit, use an overall diff operation to compute the net changed files instead of doing a "manual" diff between the full set of files and base commit.

This gets rid of last user of cmdutil.samefile() (which is a serial fetch footgun).

Reviewed By: zzl0

Differential Revision: D66203460

fbshipit-source-id: 2d38ba571fa9524c43f1c3d9a7b1bc7162140378