amend: simplify computation of new list of files
Summary:
When amending a commit, we need to compute the list of files to be included in the new commit's commit text. This is somewhat tricky since you have to be careful to remove files from the list if they were "amended away" (i.e. reverted via the amend). Previously the filtering happened essentially through a manual diff operation comparing each dirty file with the base commit. Instead we can perform a single diff between the base commit and the working copy, which already incorporates the logic of combining working copy changes with diff results.

One tweak I had to make was to allow passing a separate matcher for filtering the dirstate status vs. filtering the diff result. In this case, we don't want to filter the diff between "base" and "old", but we do want to filter the dirstate status (i.e. user's matcher should only filter which dirty files to amend).

Reviewed By: zzl0

Differential Revision: D66203461

fbshipit-source-id: 560cfabbaeedda9d6651a6d31c563c81df2bdd15