cas: avoid read-before-write for SLAPI -> tree aux cache
Summary:
When fetching trees from SLAPI and saving the tree aux data into the tree aux cache, we were reading from the tree aux cache before each write. This read is not necessary for the SLAPI case, so let's skip it.

It is necessary when fetching from CAS because we fetch trees over and over from CAS, and we don't want to keep writing the same trees into the tree aux cache over and over.

Reviewed By: liubov-dmitrieva

Differential Revision: D66147395

fbshipit-source-id: 51062ae109664bbdd8b48909e067b01071f5dc6c