apply clippy fixes
Summary:
# This diff
Applies 100s of clippy lints to the targets under `eden/scm/...`. It also applies some subsequent lints/formatting (found with `arc lint`) that were created by the fixing of the clippy lints.

# Context
I was trying to lint just a small portion of the codebase, but then I accidentally linted the entire `eden/scm/` subdirectory and decided to just go with it. There are only a few that I'm not 100% sure about, all of which I'll leave a comment by on the diff.

Reviewed By: muirdm

Differential Revision: D66340886

fbshipit-source-id: f4eca16b884bab71c1cd83c543187ae363a1a611