localrepo: set files to empty for subtree shallow copy commits
Summary: For subtree shallow copy commit, we don't traverse the tree to find all the files. So the "files" list in the commit text will be empty for pure subtree shallow copy. Further more, we want to make the "files" list empty for any commits that contains subtree shallow copy, since the "files" list will not have all the changed files if the commit contains a shallow copy.

Reviewed By: muirdm

Differential Revision: D66193902

fbshipit-source-id: 3aaf5a95533112060fd52d6c69b2e2ba0c0efe7d