Add new test for multiple Sapling requests with same nodeID in a queue
Summary:
We cannot have two requests with the same nodeID in a request queue. It will end up in completion two request promises with same results, and eden will crash with `folly::PromiseAlreadySatisfied`
The previous diff avoid sending multiple requests for same nodeID and different FetchCause. To guarantee that the previous code is working we can add this test to `SaplingBackingStoreTest` and cover this corner case.

Reviewed By: jdelliot

Differential Revision: D66337316

fbshipit-source-id: 72b3b745f97cee27c54c611601270ef3889a4f76