Avoid deduplicating backing store requests with same nodeID
Summary:
## Background
`SaplingBackingStore` prepare requests and send them to Sapling. In `prepareRequests` function, we have `importRequestsMap` to group requests by nodeID and ensure no duplicates in fetch request send to `SaplingNativeBackingStore`. However, the `prepareRequests` function also deduplicate the requests for a given nodeID based on the FetchCause.
This is wrong because we will have two requests with the same nodeID in one batch. It could fail in these lines when we want to set promise with the replied requests from backing store:
https://www.internalfb.com/code/fbsource/[be3dafab33e2]/fbcode/eden/fs/store/hg/SaplingBackingStore.cpp?lines=725%2C917
This is a very rare corner case, and it can only happen if we have a prefetch in background and exactly after that we do a fetch in foreground. In this case edenFS crash with `folly::PromiseAlreadySatisfied`
That was the root cause of S463588. On that SEV, this diff D64799979 added `background` to put a prefetch in background to speed up the process. However it ends up to eden crash and we have to revert the diff in D65158135

## repro
To repro this crash, I created a unit test in D66256624. `sameBatchRequestsDifferentFetchCause` test adds two requests with different fetch cause in a one queue and send the queue to `getTreeBacth`. As expected the test is failing because eden crash with
```
terminate called after throwing an instance of 'folly::PromiseAlreadySatisfied'
  what():  Promise already satisfied
*** Aborted at 1732157041 (Unix time, try 'date -d 1732157041') ***
*** Signal 6 (SIGABRT) (0x755900004d5c) received by PID 19804 (pthread TID 0x7f2499e00640) (linux TID 19974) (maybe from PID 19804, UID 30041) (code: -6), stack trace: ***
AddressSanitizer:DEADLYSIGNAL
```

## The fix in this diff:
The fix in this diff is avoiding deduplicating requests based on the FetchCause. FetchCause is sending to Sapling by eden. Now, we send one request for each nodeID, and only send the FetchCause with the highest priority

Reviewed By: jdelliot

Differential Revision: D66335588

fbshipit-source-id: 08b921bc39cc2f86dc4e6af164d9995681323919