Strip relative root from offloaded queries
Summary:
# Bug Description
In a recent change, we allowed certain relative roots to be allowed to
be offloaded. However, the results are still based on the server results, which are all results from the root of the repo. This issue was obscured by a different bug in watchman where extra parts from the prefix were stripped, so when that bug was fixed, this one arose.

# This Diff
Remove the relative root from the matching paths. This matches existing functionality for non-offloaded queries

# Technical Details
When collecting the results, check the prefix of the file. If there is a relative root, remove it from files that are in that path, and drop the files that aren't.

Reviewed By: jdelliot

Differential Revision: D66329531

fbshipit-source-id: b1db4d457cbeac41197180a6da108e0ccb5bd6ae