radixbuf: fix benchmark
Summary:
# This diff

Fixes the radixbuf benchmark so that it can actually be built with `cargo bench`

# Context

Clippy suggests that we change some code in the radixbuf library that does some comparisons. It was unclear to me if the original comparison was done for performance reasons. Therefore, we should benchmark the radixbuf library before/after the change to see if we should ignore the lint or apply the fix.

See next diff for the results

Reviewed By: muirdm

Differential Revision: D66375504

fbshipit-source-id: f76707b44182001eff35e48b6760e7090d0d4490