radixbuf: address clippy lint
Summary:
# This diff

Applies a clippy lint to the radixbuf library

# Context

Clippy suggests that we change some code in the radixbuf library that does some comparisons. It was unclear to me if the original comparison was done for performance reasons. Therefore, we should benchmark the radixbuf library before/after the change to see if we should ignore the lint or apply the fix.

See the test plans for the results

Reviewed By: sggutier

Differential Revision: D66375503

fbshipit-source-id: 8215986acd51a8034129536c490ee7b640baffcc