filestore: remove unused cache method
Summary:
# This diff
Removes the cache() method on the filestore.

# Context
I was cleaning up some clippy lints and I saw that the cache() method had unsafe behavior. The cache method was also writing into an uninitialized vec, which is unsafe. I also realized that the cache() method was only used in the test that's also removed in this diff, and therefore we could remove the code instead of fixing the safety.

Reviewed By: sggutier

Differential Revision: D66347615

fbshipit-source-id: d5b89ddeac1ba6de293e0f5b83bf704cbf6290eb